move the library LWRPs to custom resources
also updates for modern foodcritic and fauxhai
